Regjster.jsx functionality add
Register page now accepts new credentials and successfully adds them to MongoDB database, and when duplicate credentials are submitted a small error message appears below the register button. State management for username, password, email, and error status all added, in addition to some minor css styling. Only change in App.js is a temporary change to user status as false so that register page can be reached (unreachable when user is logged in)